# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "ShapeShift",
      "description": "OpenAPI Schema for ShapeShift",
      "version": "0.0.1"
    },
    "paths": {
...
      "/api/v1/exercise/exercises/{workout_id}": {
        "get": {
          "tags": ["exercises"],
          "summary": "Get all exercises",
          "operationId": "getExercises",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ExerciseOut"
                    },
                    "title": "Response Getexercises"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/exercise/{workout_id}": {
        "post": {
          "tags": ["exercises"],
          "summary": "Create an exercise",
          "operationId": "createExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "workout_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Workout Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Exercise"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/exercise/{exercise_id}": {
        "get": {
          "tags": ["exercises"],
          "summary": "Get a single exercise",
          "operationId": "getExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ExerciseOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["exercises"],
          "summary": "Update an exercise",
          "operationId": "updateExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ExerciseOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["exercises"],
          "summary": "Delete an exercise",
          "operationId": "deleteExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/exercise-instance/exercise-instances/exercise/{exercise_id}": {
        "get": {
          "tags": ["exercise-instances"],
          "summary": "Get all exercise instances by exercise",
          "operationId": "getExerciseInstancesByExercise",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ExerciseInstanceOut"
                    },
                    "title": "Response Getexerciseinstancesbyexercise"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/exercise-instance/exercise-instances/{session_id}": {
        "get": {
          "tags": ["exercise-instances"],
          "summary": "Get all exercise instances",
          "operationId": "getExerciseInstances",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "session_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Session Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ExerciseInstanceOut"
                    },
                    "title": "Response Getexerciseinstances"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/exercise-instance/{session_id}": {
        "post": {
          "tags": ["exercise-instances"],
          "summary": "Create an exercise instance",
          "operationId": "createExerciseInstance",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "session_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Session Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseInstanceCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ExerciseInstance"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/exercise-instance/{exercise_instance_id}": {
        "get": {
          "tags": ["exercise-instances"],
          "summary": "Get a single exercise instance",
          "operationId": "getExerciseInstance",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_instance_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Instance Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ExerciseInstanceOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["exercise-instances"],
          "summary": "Update an exercise instance",
          "operationId": "updateExerciseInstance",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_instance_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Instance Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseInstanceUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ExerciseInstanceOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["exercise-instances"],
          "summary": "Delete an exercise instance",
          "operationId": "deleteExerciseInstance",
          "security": [
            {
              "JWT": []
            }
          ],
          "parameters": [
            {
              "name": "exercise_instance_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid",
                "title": "Exercise Instance Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Exercise": {
          "properties": {
            "_id": {
              "anyOf": [
                {
                  "type": "string",
                  "example": "5eb7cf5a86d9755df3a6c593"
                },
                {
                  "type": "null"
                }
              ],
              "title": " Id",
              "description": "MongoDB document ObjectID"
            },
            "exercise_id": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id",
              "unique": true
            },
            "name": {
              "type": "string",
              "maxLength": 50,
              "minLength": 1,
              "title": "Name"
            },
            "type": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "equipment": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "muscle": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Muscle"
            },
            "track_reps": {
              "type": "boolean",
              "title": "Track Reps",
              "default": false
            },
            "track_weight": {
              "type": "boolean",
              "title": "Track Weight",
              "default": false
            },
            "track_distance": {
              "type": "boolean",
              "title": "Track Distance",
              "default": false
            },
            "track_duration": {
              "type": "boolean",
              "title": "Track Duration",
              "default": false
            },
            "workout": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Workout"
            },
            "owner": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Owner"
            }
          },
          "type": "object",
          "required": ["name", "workout", "owner"],
          "title": "Exercise"
        },
        "ExerciseCreate": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "type": {
              "$ref": "#/components/schemas/ExerciseType"
            },
            "muscle": {
              "type": "string",
              "title": "Muscle"
            },
            "equipment": {
              "type": "string",
              "title": "Equipment"
            },
            "description": {
              "type": "string",
              "title": "Description"
            }
          },
          "type": "object",
          "required": ["name", "type", "muscle", "equipment", "description"],
          "title": "ExerciseCreate"
        },
...
        "ExerciseInstance": {
          "properties": {
            "_id": {
              "anyOf": [
                {
                  "type": "string",
                  "example": "5eb7cf5a86d9755df3a6c593"
                },
                {
                  "type": "null"
                }
              ],
              "title": " Id",
              "description": "MongoDB document ObjectID"
            },
            "exercise_instance_id": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Instance Id",
              "unique": true
            },
            "sets": {
              "type": "integer",
              "title": "Sets",
              "default": 1
            },
            "reps": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Reps",
              "default": []
            },
            "weight": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Weight",
              "default": []
            },
            "distance": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Distance",
              "default": []
            },
            "duration": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Duration",
              "default": []
            },
            "exercise": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Exercise"
            },
            "session": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Session"
            },
            "owner": {
              "properties": {
                "id": {
                  "type": "string",
                  "title": "Id"
                },
                "collection": {
                  "type": "string",
                  "title": "Collection"
                }
              },
              "type": "object",
              "required": ["id", "collection"],
              "title": "Owner"
            }
          },
          "type": "object",
          "required": ["exercise", "session", "owner"],
          "title": "ExerciseInstance"
        },
        "ExerciseInstanceCreate": {
          "properties": {
            "exercise_id": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          },
          "type": "object",
          "required": ["exercise_id"],
          "title": "ExerciseInstanceCreate"
        },
        "ExerciseInstanceOut": {
          "properties": {
            "exercise_instance_id": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Instance Id"
            },
            "sets": {
              "type": "integer",
              "title": "Sets"
            },
            "reps": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Reps"
            },
            "weight": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Weight"
            },
            "distance": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Distance"
            },
            "duration": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Duration"
            },
            "exercise": {
              "$ref": "#/components/schemas/Exercise"
            },
            "session": {
              "$ref": "#/components/schemas/Session"
            }
          },
          "type": "object",
          "required": [
            "exercise_instance_id",
            "sets",
            "reps",
            "weight",
            "distance",
            "duration",
            "exercise",
            "session"
          ],
          "title": "ExerciseInstanceOut"
        },
        "ExerciseInstanceUpdate": {
          "properties": {
            "sets": {
              "type": "integer",
              "title": "Sets"
            },
            "reps": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Reps"
            },
            "weight": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Weight"
            },
            "distance": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Distance"
            },
            "duration": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Duration"
            }
          },
          "type": "object",
          "required": ["sets", "reps", "weight", "distance", "duration"],
          "title": "ExerciseInstanceUpdate"
        },
        "ExerciseOut": {
          "properties": {
            "exercise_id": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            },
            "name": {
              "type": "string",
              "title": "Name"
            },
            "type": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "muscle": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Muscle"
            },
            "equipment": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "track_reps": {
              "type": "boolean",
              "title": "Track Reps"
            },
            "track_weight": {
              "type": "boolean",
              "title": "Track Weight"
            },
            "track_distance": {
              "type": "boolean",
              "title": "Track Distance"
            },
            "track_duration": {
              "type": "boolean",
              "title": "Track Duration"
            }
          },
          "type": "object",
          "required": [
            "exercise_id",
            "name",
            "type",
            "muscle",
            "equipment",
            "description",
            "track_reps",
            "track_weight",
            "track_distance",
            "track_duration"
          ],
          "title": "ExerciseOut"
        },
        "ExerciseType": {
          "type": "string",
          "enum": [
            "cardio",
            "strength",
            "olympic_weightlifting",
            "plyometrics",
            "powerlifting",
            "stretching",
            "strongman"
          ],
          "title": "ExerciseType"
        },
        "ExerciseUpdate": {
          "properties": {
            "name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            },
            "type": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExerciseType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "muscle": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Muscle"
            },
            "equipment": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "track_reps": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Reps"
            },
            "track_weight": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Weight"
            },
            "track_distance": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Distance"
            },
            "track_duration": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Track Duration"
            }
          },
          "type": "object",
          "required": [
            "name",
            "type",
            "muscle",
            "equipment",
            "description",
            "track_reps",
            "track_weight",
            "track_distance",
            "track_duration"
          ],
          "title": "ExerciseUpdate"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": ["loc", "msg", "type"],
          "title": "ValidationError"
        },
...
      },
      "securitySchemes": {
        "JWT": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {},
              "tokenUrl": "/api/v1/auth/login"
            }
          }
        }
      }
    },
    "servers": [
      {
        "url": "https://shapeshift-backend.onrender.com"
      }
    ]
  }